+++ steel/sources/res/text/text_file.cpp
@@ -36,7 +36,7 @@ bool TextFile::init(const std::string& name, const std::string dir)
 
 	pbuf = f.rdbuf();
 	// get file size using buffer's members
-	size = pbuf->pubseekoff (0,ios::end,ios::in);
+	size = (long)pbuf->pubseekoff (0,ios::end,ios::in);
 	length = size;
 	pbuf->pubseekpos (0,ios::in);
 	// allocate memory to contain file data
@@ -67,7 +67,7 @@ char* readfile(std::string file)
 
 	pbuf = f.rdbuf();
 	// get file size using buffer's members
-	size = pbuf->pubseekoff (0,ios::end,ios::in);
+	size = (long)pbuf->pubseekoff (0,ios::end,ios::in);
 	pbuf->pubseekpos (0,ios::in);
 	// allocate memory to contain file data
 	text = new char[size + 1];
